Refactor codes using cleared tyled components
