+++ src/App.jsx
@@ -4,6 +4,7 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Data from './components/data/Data';
 import MainPage from './components/home/mainPage';
 import Map from './components/map/Map';
+import AdminMap from './components/managementUI/map/AdminMap';
 import SavingsGroupsdata from './components/data/SavingsCreditData';
 import ConsentFormData from './components/data/Data';
 import XlsxUpload from './components/uploadFiles/UploadData';
@@ -15,6 +16,8 @@ import ConfirmResetPass from './components/authentication/resetPassword/ConfirmR
 import Layout from './components/helpers/Layout';
 import LoginOTP from './components/authentication/singin/loginOtp';
 import { ToastContainer } from 'react-toastify';
+import ManagementData from './components/managementUI/admindata/ManagementData';
+import NgosManagement from './components/managementUI/admindata/NgosManagement';
 
 const App = () => {
     return (
@@ -36,6 +39,12 @@ const App = () => {
                         element={<ConfirmResetPass />}
                     ></Route>
                     <Route path="/activate" element={<LoginOTP />}></Route>
+                    <Route path="/savings-group-map/admin" element={<AdminMap />}></Route>
+                    <Route
+                        path="/savings-group-map/admin/management"
+                        element={<ManagementData />}
+                    ></Route>
+                    <Route path="/savings-group-map/admin/management/data" element={<NgosManagement />}></Route>
                 </Routes>
             </Layout>
         </Router>
