+++ src/components/managementUI/map/AdminMapNavbar.jsx
@@ -1,15 +1,20 @@
 import React from 'react';
 import styled from 'styled-components';
-import picture from '../../assets/bnr.svg';
-import profile from '../../assets/Profile.png';
+import picture from '../../../assets/bnr.svg';
+import profile from '../../../assets/Profile.png';
 import { Link } from 'react-router-dom';
+// import CountryDropDown from './CountryDropDown';
 
 const Container = styled.div`
     height: 85px;
-    box-shadow: 0px 4px 4px 0px #00000008;
     display: flex;
     justify-content: space-between;
     width: 100%;
+    background-color: white;
+    box-shadow: 0px 4px 4px 0px #00000008;
+    overflow: hidden;
+    /* position: fixed; */
+    z-index: 999;
 `;
 const Wrapper = styled.div`
     padding: 10px 20px;
@@ -19,7 +24,7 @@ const Wrapper = styled.div`
 `;
 
 const Left = styled.div`
-    width: 50%;
+    width: 60%;
     display: flex;
     align-items: center;
     justify-content: space-around;
@@ -50,6 +55,24 @@ const Menu = styled.ul`
     }
 `;
 
+const InputFieldSelect = styled.select`
+    width: 60%;
+    background-color: #f2f2f2;
+    display: flex;
+    flex-direction: row;
+    align-items: flex-start;
+    padding: 3%;
+    border: 1.2px solid #dee1e7;
+    border-radius: 4px;
+    text-align: center;
+    margin-left: -38%;
+
+    ::placeholder {
+        color: black;
+        font-size: 12px;
+    }
+`;
+
 const MenuItem = styled.li`
     margin-right: 29px;
     font-size: 15px;
@@ -59,12 +82,16 @@ const MenuItem = styled.li`
 `;
 
 const Button = styled.button`
-    padding: 2px 3px;
+    padding: 2%;
     color: black;
     font-weight: 500;
     cursor: pointer;
-    font-size: 15px;
-    margin-right: 24px;
+    font-size: 14px;
+    margin-right: 3%;
+    border: none;
+    width: 60%;
+    align-items: center;
+    background-color: white;
 `;
 
 const Picture = styled.img`
@@ -74,11 +101,18 @@ const Picture = styled.img`
 `;
 
 const Profile = styled.img`
-    width: 20px;
-    height: 20px;
+    width: 1.2rem;
+    height: 1.3rem;
     object-fit: cover;
-    margin-bottom: 4px;
     cursor: pointer;
+    position: absolute;
+    margin-left: 4%;
+`;
+
+const InputOption = styled.option`
+    background-color: red;
+    color: white;
+    border: solid 1px red;
 `;
 
 const StyledLink = styled(Link)`
@@ -98,7 +132,7 @@ const StyledLink = styled(Link)`
     }
 `;
 
-const DataNavbar = () => {
+const AdminMapNavbar = () => {
     return (
         <Container>
             <Left>
@@ -118,6 +152,14 @@ const DataNavbar = () => {
                         </StyledLink>
                     </MenuItem>
                     <MenuItem>
+                        <StyledLink
+                            className="animate__animated"
+                            to="/savings-group-map/admin/management"
+                        >
+                            NGO Management
+                        </StyledLink>
+                    </MenuItem>
+                    <MenuItem>
                         <StyledLink className="animate__animated" to="/savings-group-map/data">
                             Data
                         </StyledLink>
@@ -125,11 +167,26 @@ const DataNavbar = () => {
                 </Menu>
             </Left>
             <Wrapper>
+                <InputFieldSelect
+                    classname="js-example-basic-single"
+                    id="0"
+                    name="country"
+                    type="text"
+                    placeholder="Year :"
+                    className="inputs"
+                >
+                    <InputOption value="2022">Year :2022</InputOption>
+                    <InputOption value="2021">Year :2021</InputOption>
+                    <InputOption value="2020">Year :2020</InputOption>
+                    <InputOption value="2019">Year :2019</InputOption>
+                    <InputOption value="2018">Year :2018</InputOption>
+                    <InputOption value="2017">Year :2017</InputOption>
+                </InputFieldSelect>
                 <Profile src={profile} />
-
+                <Button>Mugisha Roger</Button>
             </Wrapper>
         </Container>
     );
 };
 
-export default DataNavbar;
+export default AdminMapNavbar;
